Add pass updates

+++ app/src/main/java/nz/eloque/foss_wallet/persistence/pass/PassDao.kt
@@ -18,7 +18,7 @@ interface PassDao {
 
     @Transaction
     @Query("SELECT * FROM pass WHERE id=:id")
-    fun byId(id: Int): PassWithLocalization
+    fun byId(id: Long): PassWithLocalization
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun insert(pass: Pass): Long
