+++ app/src/main/java/nz/eloque/foss_wallet/persistence/pass/PassRepository.kt
@@ -8,7 +8,7 @@ import nz.eloque.foss_wallet.persistence.PassBitmaps
 interface PassRepository {
     fun all(): Flow<List<PassWithLocalization>>
 
-    suspend fun byId(id: Int): PassWithLocalization
+    suspend fun byId(id: Long): PassWithLocalization
 
     suspend fun insert(data: Pair<Pass, PassBitmaps>): Long
